feat(common): make CompletionQueueImpl mockable [2]
Made `CompletionQueueImpl` a pure abstract class, that makes it fully
mockable. Move the implementation to `DefaultCompletionQueueImpl`. I had
to move some other classes around to compile the code and avoid strange
helpers in `completion_queue_impl.h`.
