+++ google/cloud/completion_queue_test.cc
@@ -14,6 +14,7 @@
 
 #include "google/cloud/completion_queue.h"
 #include "google/cloud/future.h"
+#include "google/cloud/internal/default_completion_queue_impl.h"
 #include "google/cloud/testing_util/assert_ok.h"
 #include "google/cloud/testing_util/fake_completion_queue_impl.h"
 #include <google/bigtable/admin/v2/bigtable_table_admin.grpc.pb.h>
@@ -568,7 +569,7 @@ TEST(CompletionQueueTest, CancelTimerContinuation) {
 }
 
 TEST(CompletionQueueTest, ImplStartOperationDuplicate) {
-  auto impl = std::make_shared<internal::CompletionQueueImpl>();
+  auto impl = std::make_shared<internal::DefaultCompletionQueueImpl>();
   auto op = std::make_shared<MockOperation>();
 
   impl->StartOperation(op, [](void*) {});
