+++ google/cloud/google_cloud_cpp_grpc_utils.bzl
@@ -29,8 +29,10 @@ google_cloud_cpp_grpc_utils_hdrs = [
     "internal/async_read_stream_impl.h",
     "internal/async_retry_loop.h",
     "internal/async_retry_unary_rpc.h",
+    "internal/async_rpc_details.h",
     "internal/background_threads_impl.h",
     "internal/completion_queue_impl.h",
+    "internal/default_completion_queue_impl.h",
     "internal/log_wrapper.h",
     "internal/pagination_range.h",
     "internal/retry_loop.h",
@@ -44,6 +46,7 @@ google_cloud_cpp_grpc_utils_srcs = [
     "grpc_error_delegate.cc",
     "internal/background_threads_impl.cc",
     "internal/completion_queue_impl.cc",
+    "internal/default_completion_queue_impl.cc",
     "internal/log_wrapper.cc",
     "internal/retry_loop_helpers.cc",
     "internal/time_utils.cc",
